Enum type conversion and member value access.
- Added tests for the type conversion part.

- Enum member value access still needs some work

+++ SolidityEndToEndTest.cpp
@@ -2510,7 +2510,7 @@ BOOST_AUTO_TEST_CASE(using_enums)
 				}
 				function getChoice() returns (uint d)
 				{
-					d = choices;
+					d = uint256(choices);
 				}
 				ActionChoices choices;
 			}
